# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KwunMan OS - 冠文鋼結構管理系統 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects with optional filters
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: ["報價中", "待通知", "待收首期", "施工中", "待收款", "待收尾期", "已完成", "不成功"]
        - name: taxView
          in: query
          schema:
            type: string
            enum: ["kwunman", "gov"]
        - name: quarter
          in: query
          schema:
            type: string
        - name: client
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project / quote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
  /projects/quarterly-stats:
    get:
      operationId: getQuarterlyStats
      tags: [projects]
      summary: Get quarterly financial statistics
      parameters:
        - name: taxView
          in: query
          required: true
          schema:
            type: string
            enum: ["kwunman", "gov"]
      responses:
        "200":
          description: Quarterly statistics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuarterlyStat"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a single project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/mark-received:
    post:
      operationId: markProjectReceived
      tags: [projects]
      summary: Mark project balance as fully received
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                receivedDate:
                  type: ["string", "null"]
                finalReceived:
                  type: ["number", "null"]
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  # ─── INVOICES ─────────────────────────────────────────────────────────────────
...
  /quote-items:
    get:
      operationId: listQuoteItems
      tags: [projects]
      summary: List line items for a quote
      parameters:
        - name: quoteId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of quote items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuoteItem"
    post:
      operationId: createQuoteItem
      tags: [projects]
      summary: Add a line item to a quote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteItemInput"
      responses:
        "201":
          description: Created quote item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteItem"
  /quote-items/{itemId}:
    patch:
      operationId: updateQuoteItem
      tags: [projects]
      summary: Update a line item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteItemUpdate"
      responses:
        "200":
          description: Updated quote item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteItem"
    delete:
      operationId: deleteQuoteItem
      tags: [projects]
      summary: Delete a line item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── CLIENTS ──────────────────────────────────────────────────────────────────
...
  /payslips:
    get:
      operationId: listPayslips
      tags: [payslips]
      summary: List payslips
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of payslips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payslip"
    post:
      operationId: createPayslip
      tags: [payslips]
      summary: Create a payslip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayslipInput"
      responses:
        "201":
          description: Created payslip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payslip"
  /payslips/{payslipId}:
    get:
      operationId: getPayslip
      tags: [payslips]
      summary: Get a specific payslip
      parameters:
        - name: payslipId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payslip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payslip"
    delete:
      operationId: deletePayslip
      tags: [payslips]
      summary: Delete a payslip
      parameters:
        - name: payslipId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Project:
      type: object
      required:
        - id
        - quoteId
        - date
        - location
        - projectItem
        - quoteAmount
        - client
        - clientId
        - status
        - depositStatus
        - startDate
        - endDate
        - invoiceStatus
        - finalReceived
        - balanceStatus
        - taxQuarterKwunman
        - taxQuarterGov
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        quoteId:
          type: string
          description: Quote number e.g. "Q001"
        date:
          type: ["string", "null"]
          description: Quote issue date YYYY-MM-DD
        location:
          type: string
        projectItem:
          type: string
        quoteAmount:
          type: number
        client:
          type: string
        clientId:
          type: string
        status:
          type: string
          enum: ["報價中", "待通知", "待收首期", "施工中", "待收款", "待收尾期", "已完成", "不成功"]
        depositStatus:
          type: string
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        invoiceStatus:
          type: string
        finalReceived:
          type: number
        balanceStatus:
          type: string
          enum: ["待收尾數", "已收齊", "無"]
        taxQuarterKwunman:
          type: string
        taxQuarterGov:
          type: string
        projectType:
          type: ["string", "null"]
        paymentTerms:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        totalPaid:
          type: number
          description: Sum of all payment entries (computed)
        createdAt:
          type: string
        updatedAt:
          type: string
    ProjectInput:
      type: object
      required:
        - location
        - projectItem
        - quoteAmount
        - client
        - clientId
        - status
      properties:
        quoteId:
          type: ["string", "null"]
        date:
          type: ["string", "null"]
        location:
          type: string
        projectItem:
          type: string
        quoteAmount:
          type: number
        client:
          type: string
        clientId:
          type: string
        status:
          type: string
          enum: ["報價中", "待通知", "待收首期", "施工中", "待收款", "待收尾期", "已完成", "不成功"]
        depositStatus:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        invoiceStatus:
          type: ["string", "null"]
        finalReceived:
          type: ["number", "null"]
        balanceStatus:
          type: ["string", "null"]
        projectType:
          type: ["string", "null"]
        paymentTerms:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    ProjectUpdate:
      type: object
      properties:
        date:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        projectItem:
          type: ["string", "null"]
        quoteAmount:
          type: ["number", "null"]
        client:
          type: ["string", "null"]
        clientId:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
          enum: ["報價中", "待通知", "待收首期", "施工中", "待收款", "待收尾期", "已完成", "不成功", null]
        depositStatus:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        invoiceStatus:
          type: ["string", "null"]
        finalReceived:
          type: ["number", "null"]
        balanceStatus:
          type: ["string", "null"]
          enum: ["待收尾數", "已收齊", "無", null]
        projectType:
          type: ["string", "null"]
        paymentTerms:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...
    QuarterlyStat:
      type: object
      required:
        - quarter
        - totalReceived
        - totalQuoted
        - projectCount
        - completedCount
      properties:
        quarter:
          type: string
        totalReceived:
          type: number
        totalQuoted:
          type: number
        projectCount:
          type: integer
        completedCount:
          type: integer
...
    QuoteItem:
      type: object
      required:
        - id
        - quoteId
        - seqNo
        - description
        - unitPrice
        - qty
        - createdAt
      properties:
        id:
          type: integer
        quoteId:
          type: string
        seqNo:
          type: integer
        description:
          type: string
        unitPrice:
          type: number
        qty:
          type: number
        notes:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    QuoteItemInput:
      type: object
      required:
        - quoteId
        - seqNo
        - description
        - unitPrice
        - qty
      properties:
        quoteId:
          type: string
        seqNo:
          type: integer
        description:
          type: string
        unitPrice:
          type: number
        qty:
          type: number
        notes:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
    QuoteItemUpdate:
      type: object
      properties:
        seqNo:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        unitPrice:
          type: ["number", "null"]
        qty:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]

    # ─── CLIENT ───────────────────────────────────────────────────────────────
...
    Payslip:
      type: object
      required:
        - id
        - employeeId
        - year
        - month
        - basicSalary
        - allowances
        - overtime
        - deductions
        - netPay
        - issuedAt
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        year:
          type: integer
        month:
          type: integer
        basicSalary:
          type: number
        allowances:
          type: number
        overtime:
          type: number
        deductions:
          type: number
        netPay:
          type: number
        notes:
          type: ["string", "null"]
        issuedAt:
          type: string
...
    PayslipInput:
      type: object
      required:
        - employeeId
        - year
        - month
        - basicSalary
        - netPay
      properties:
        employeeId:
          type: integer
        year:
          type: integer
        month:
          type: integer
        basicSalary:
          type: number
        allowances:
          type: number
        overtime:
          type: number
        deductions:
          type: number
        netPay:
          type: number
        notes:
          type: ["string", "null"]